+++ src/Main.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import App from "./components/App";
-import ItemPage from "./pages/ItemPage";
-import HomePage from "./pages/HomePage";
+import ItemPage from "./pages/ItemPage/ItemPage";
+import HomePage from "./pages/HomePage/HomePage";
 
 function Main() {
   return (
